dont use an extension for Interfaces.jl
