+++ src/DimensionalData.jl
@@ -24,7 +24,7 @@ import Adapt,
        ArrayInterface,
        ConstructionBase, 
        Extents,
-       InterfacesCore,
+       Interfaces,
        InvertedIndices,
        IteratorInterfaceExtensions,
        RecipesBase,
