+++ src/array/array.jl
@@ -19,9 +19,6 @@ on the `Dimension` index. Use forward-ordered arrays only"
 """
 abstract type AbstractDimArray{T,N,D<:Tuple,A} <: AbstractArray{T,N} end
 
-InterfacesCore.@interface_core DimArrayInterface AbstractDimArray
-
-
 const AbstractDimVector = AbstractDimArray{T,1} where T
 const AbstractDimMatrix = AbstractDimArray{T,2} where T
 const AbstractDimVecOrMat = Union{AbstractDimVector,AbstractDimMatrix}
