+++ src/stack/stack.jl
@@ -27,8 +27,6 @@ The constructor of an `AbstractDimStack` must accept a `NamedTuple`.
 """
 abstract type AbstractDimStack{L} end
 
-InterfacesCore.@interface_core DimStackInterface AbstractDimStack
-
 data(s::AbstractDimStack) = getfield(s, :data)
 dims(s::AbstractDimStack) = getfield(s, :dims)
 refdims(s::AbstractDimStack) = getfield(s, :refdims)
